% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compiled_data_json.R
\name{compiled_data_json}
\alias{compiled_data_json}
\title{compiled_data_json: Generates a json file with descriptions of demographic and dexa variables}
\usage{
compiled_data_json()
}
\value{
A string with data stored in JSON format containing meta-data for the TSV file exported by code/compile_databases.R
}
\description{
This function generates a json file for the TSV file exported by code/compile_databases.R
}
