% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataBRAKEr.R
\name{dataBRAKEr}
\alias{dataBRAKEr}
\title{dataBRAKEr: Process Study BRAKE data}
\usage{
dataBRAKEr(
  base_wd,
  overwrite = FALSE,
  fnirs_overwrite = FALSE,
  proc_source = FALSE,
  data_list = "all",
  data_type = "all",
  micro_protocols = "all",
  micro_data_type = "all",
  proc_ggir = FALSE,
  overwrite_ggir_derivs = FALSE,
  return_data = FALSE,
  deid = FALSE,
  deid_list = "all",
  openneuro_org = FALSE
)
}
\arguments{
\item{base_wd}{full path to directory containing both raw_untouched and bids directories}

\item{overwrite}{logical indicating if data should be overwritten. Default = FALSE}

\item{fnirs_overwrite}{overwrite fNIRS data files - separate argument because the Matlab fNIRS processing writes onset values to the rawdata/*_events.tsv values. Overwriting will overwrite these values with NA.}

\item{proc_source}{whether to processes the raw data using proc_tasks.R. Default = FALSE}

\item{data_list}{list of strings matching the notes below to indicate the data to be written. Default = 'all' to export all data and metadata. Options include:
\itemize{
\item{'participants' - BIDS specified participants.tsv file}
  \itemize{tasks:
   \item{'foodrating' - fNIRS Food Rating task}
   \item{'foodchoice' - fNIRS Food Choice task}
   \item{'shapegame' - Shape Game data}
   \item{'spacegame' - Space Game data (need to finish processing in Matlab)}
   \item{'nihtoolbox' - NIH Toolbox data}
   \item{'tastetest' - fNIRS Taste-Test task}
   \item{'pit' - Pavlovian Instrumental Transfer task data}
 }
 \item{'actigraph' - activity and sleep data generated from GGIR and mMARCH.AC}
 \item{'microstructure' - coded meal microstructure data}
 \item{'paticipants' - BIDS specified participants.tsv file}
 \item{'anthropometrics' - height, weight, and computed anthropometric data}
 \item{'demographics' - compiled demographic data}
 \item{'bodpod' - verified BodPod data}
 \item{'fnirs_info' - compiled fNIRS task-related information}
 \item{'dkefs' - verified D-KEFS data}
 \item{'wasi' - verified WASI data}
 \item{'intake' - compiled verified intake data with computed intake values}
 \item{'tasttest_samples' - verified Taste-Test sample weights}
 \item{'household' - compiled demographicinformation about houshold}
 \item{'infancy' - compiled demographic information related to infancy}
 \item{'bes' - Binge Eating Scale}
 \item{'brief2' - Behavioral Rating Inventory of Executive Function-2}
 \item{'cbq' - Child Behavior Questionnaire}
 \item{'cebq' - Children's Eating Behavior Questionnaire}
 \item{'cfq' - Child Feeding Questionnaire}
 \item{'cshq' - Children Sleep Habits Questionnaire}
 \item{'cwc' - Child Weight Concerns Questionnaire}
 \item{'efcr' - External Food Cue Responsiveness Scale}
 \item{'ffbs' - Family Food Behavior Survey}
 \item{'ffq' - HELIX cohort Food Frequency Questionnaire}
 \item{'fmcb' - Feeding to Manage Child Behavior Questionnaire}
 \item{'hfe' - Home Food Environment}
 \item{'hfi' - Fulkerson Home Food Inventory}
 \item{'lbc' - Lifestyle Behavior Checklist}
 \item{'loc' - Loss of Control-Eating Questionnaire}
 \item{'puberty' - combination of Tanner and Pubertal Rating Scale}
 \item{'pwlb' - Parent Weight-Loss Behavior Questionnaire}
 \item{'scpf' - Structure and Control of Parent Feeding Questionnaire}
 \item{'sic' - Stress in Children Questionnaire}
 \item{'sleeplog' - Week long sleep log}
 \item{'spsrq' - Sensitivity to Punishment and Sensitivity to Reward Questionnaire}
 \item{'tfeq' - Three Factor Eating Questionnaire}
}}

\item{micro_protocols}{(optional) list of strings matching the notes below to indicate the which microstructure data. Default = 'all' to export all data and metadata. Options include:
\itemize{
 \item{'meal-baseline' - meal microstructure behavior at baseline}
 \item{'meal-followup' - meal microstructure behavior at followup}
 \item{'eah' - EAH microstructure behavior}
}}

\item{micro_data_type}{(optional) Type of data to process for meal microstructure - list of strings matching the data types listed below. Default = 'all' to export both:
\itemize{
  \item{'beh_wide' - summary behavioral measures in wide formate by coder. Note: this will write out a summary dataset in bids/phenotype.}
  \item{'events_long' - event level data in log format by coder. Note: this writes out a file per participant into bids/rawdata.}
}}

\item{proc_ggir}{(logical) run the GGIR and mMARCH processing step. Default = FALSE}

\item{overwrite_ggir_derivs}{(required if 'actigraph' is in data_list or data_list = 'all') overwrite GGIR derivatives. Defult = FLASE, this will take a LONG time.}

\item{deid}{(logical) generate de-identified data into scholar-sphere directory. Do not include data_list argument to run this funciton.}

\item{deid_list}{data to process. Options include 'all' to process all data that will be put on ScholarSphere or a list of the following:
\itemize{
 \item{'phenotype' - all phenotype informaiton}
 \item{'derivatives' - all derivative and processed individual task data}
}}

\item{openneuro_org}{(logical) organize and copy over fNIRS data for open-neuro directory organizaiton and upload}
}
\value{
Does not return anything
}
\description{
This function:
\itemize{Calls core proc_\* and write_\* functions and writes BIDS-compliant .tsv and .json files into bids/phenotype:
   \item{write_redcap: processes and writes out specified REDCap data}
   \item{proc_microstructure: processes and writes out microstructure data}
   \item{proc_tasks:}
}
}
\details{
To use this function, the correct path must be used. The path must be the full path to the data file, including the file name.
}
\examples{

\dontrun{
dataBRAKEr(base_wd, overwrite = FALSE, data_list = 'all')

}

}
