% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_brake_db.R
\name{gen_BRAKE_db}
\alias{gen_BRAKE_db}
\title{gen_BRAKE_db: Generate databases for the Study of Brain, Reward, and Kids' Eating (BRAKE)}
\usage{
gen_BRAKE_db(
  data_elements,
  write_dat = TRUE,
  write_path,
  data_path,
  return_data = FALSE
)
}
\arguments{
\item{data_elements}{list of strings to indicate which data to include in the database. All strings should be entered as a list. example: c('demo',' cebq', 'cbq'). Options include the following (must match string exactly):
\itemize{
 \item{Questionnaires:}
 \itemize{
   \item{bes - Binge Eating Scale}
   \item{brief2 - Behavioral Rating Inventory of Executive Functioning}
   \item{cams - Child Anxiety Meter Scale}
   \item{cbis - Child Body Image Scale}
   \item{cbq - Child Behavior Questionnaire}
   \item{cebq - Children's Eating Behavior Questionnaire}
   \item{cfq - Child Feeding Questionnaire}
   \item{cshq - Childrens Sleep Habits Questionnaire}
   \item{demo - all demographics related information}
   \item{efcr - External Food Cue Responsivenes Scale}
   \item{ffbs - Family Food Behavior Survey}
   \item{ffq_helix - Food Frequency Questionniare from the HELIX cohort study}
   \item{hmcb - Feeding to Manage Child Behavior}
   \item{hfi - Home Food Inventory}
   \item{lbc - Lifestyle Behavior Checklist}
   \item{loc - Loss of Control-Eating Questionnaire}
   \item{nik_hfe - Home Food Environment Survey from the Neighborhood Impact on Kids}
   \item{pds - Pubertal Development Scale}
   \item{sic - Stress in Children Questionnaire}
   \item{sleeplog - Week Sleep Log}
   \item{spsrq - Sensitivity to Punishment Sensitivity to Reward Questionnaire}
   \item{tfeq - Three Factor Eating Questionnaire}
   \item{wic - Weight Concerns Scale}
 }
 \item{Tasks:}
 \itemize{
   \item{anthro - Measured height and weight, BMI, and CDC BMI percentiles}
   \item{dkefs - Behavioral data for the D-KEFS Trail Making, Verbal Fluency, and Design Fluency}
   \item{foodchoice - Pre-process and produce behavioral data from the Food Choice fNIRS Task}
   \item{foodrating - Pre-process and produce behavioral data from the Food Rating fNIRS Task}
   \item{nihtoolbox - Behavioral data for NIH Toolbox tasks: Listsort, Flanker, and Dimensional Card Sort Tasks}
   \item{pit - Behavioral data for PIT task}
   \item{shapegame - Pre-process and produce behavioral and eye-tracking data for the VMAC Shape Game}
   \item{spacegame - Behavioral data for the Space Game}
   \itemize{
     \item{DM Model not processed: will produce only data needed to process task in Matlab to get DM model parameters}
     \item{DM Model exisits: will produce a database with behavioral and process DM model parameters}
     }
   \item{tasttest - Pre-process and produce behavioral data from the Food Choice fNIRS Task}
   \item{wasi - WASI IQ scores}
   }
 \item{Protocol Elements:}
 \itemize{
   \item{actigraph: Activity and sleep data}
   \item{eah: Intake and fullness data for EAH}
   \item{eah_micro: EAH behavioral coding}
   \item{meal: Intake and fullness data for meal}
   \item{meal_micro: Mictrostructure data for meal}
   \item{meal_CIC: Cumultaive intake curves for meal}
   \item{bodpod: BodPod metrics}
 }
}
\cr}

\item{write_dat}{indicate whether to write databases. Default is TRUE.}

\item{write_path}{(optional) a string with the path indicating where to save the generated databases if write_dat is TRUE (default option). If no path is given, databases will be written to working directory.}

\item{return_data}{indicate whether to return databases to the environment. Default is FALSE.}
}
\value{
A list containing all databases that were generated
}
\description{
This function will generate up-to-date databases for the Study of Brain, Reward, and Kids' Eating (BRAKE).
}
\examples{
#if in same working directory as data, want all databases, and the databases to be written to working directory:
fbs_data_proc <- gen_BRAKE_db(model_DD = TRUE)

#if only want the Demographics and Intake databases written to working directory:
fbs_data_proc <- gen_BRAKE_db(databases = c('demo', 'intake'))

#if want to work with the Demographics database but not write it to a file:
fbs_data_proc <- gen_BRAKE_db(databases = 'demo', write_dat = FALSE)

\dontrun{
#databases must be strings. The following will not run:
fbs_data_proc <- gen_BRAKE_db(databases = c(demo, intake))

}

}
\seealso{
Raw data from Qualtrics is processed using the following scripts: \code{\link{util_fbs_merge_v1dat}}, \code{\link{util_fbs_merge_v2dat}}, \code{\link{util_fbs_merge_v3dat}}, \code{\link{util_fbs_merge_v4dat}}, \code{\link{util_fbs_merge_v5dat}}, \code{\link{util_fbs_merge_v6dat}}, \code{\link{util_fbs_merge_v7dat}}
}
