% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_bes.R, R/motion_json.R
\name{json_bes}
\alias{json_bes}
\title{json_bes: Generates a json file for the Binge Eating Scale}
\usage{
json_bes()

json_bes()
}
\value{
A json file documenting the raw inputs and scored values for the Binge Eating Scale

A json file documenting the raw inputs and scored values for the Binge Eating Scale
}
\description{
This function generates a json file for the scored Binge Eating Scale and raw participant responses. This function provides accurate json files ONLY if data is processed using score_bes function in dataprepr and is only accurate for data collected in Study BRAKE.

This function generates a json file for the scored Binge Eating Scale and raw participant responses. This function provides accurate json files ONLY if data is processed using score_bes function in dataprepr and is only accurate for data collected in Study BRAKE.
}
