% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_bodpod.R
\name{json_bodpod}
\alias{json_bodpod}
\title{json_bodpod: Generates a json file for BodPod}
\usage{
json_bodpod()
}
\value{
A json file documenting the Study BRAKE's BodPod
}
\description{
This function generates a json file for BodPod
}
