% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_body_es.R
\name{json_body_es}
\alias{json_body_es}
\title{json_body_es: Generates a json file for the Body Esteem Scale}
\usage{
json_body_es()
}
\value{
A string with data stored in JSON format containing meta-data for the Body Esteem Scale
}
\description{
This function generates a json file for the scored Body Esteem Scale and raw participant responses.
}
