% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_cams.R
\name{json_cams}
\alias{json_cams}
\title{json_cams: Generates a json file for Child Anxiety Meter scale data from double entry REDCap}
\usage{
json_cams()
}
\value{
A string with data stored in JSON format containing meta-data for Child Anxiety Meter scale data
}
\description{
This function generates a json file for Child Anxiety Meter scale data
}
