% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_child_anthro.R
\name{json_child_anthro}
\alias{json_child_anthro}
\title{json_child_anthro: Generates a json file for anthropometrics recorded in REDCap}
\usage{
json_child_anthro()
}
\value{
A json file documenting the Study BRAKE anthropometrics recorded in REDCap
}
\description{
This function generates a json file for anthropometrics recorded in REDCap
}
