% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_cwc.R
\name{json_cwc}
\alias{json_cwc}
\title{json_cwc: Generates a json file for the Child Weight Concerns scale}
\usage{
json_cwc()
}
\value{
A string with data stored in JSON format containing meta-data for the Child Weight Concerns scale
}
\description{
This function generates a json file for the scored Child Weight Concerns scale and raw participant responses.
}
