% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_demo.R
\name{json_demo}
\alias{json_demo}
\title{json_demo: Generates a json file for demographic data}
\usage{
json_demo()
}
\value{
A json file documenting the Study BRAKE demographic data
}
\description{
This function generates a json file for demographic data
}
