% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_dkefs.R
\name{json_dkefs}
\alias{json_dkefs}
\title{json_dkefs: Generates a json file for D-KEFS}
\usage{
json_dkefs()
}
\value{
A json file documenting the Study BRAKE's D-KEFS
}
\description{
This function generates a json file for D-KEFS
}
