% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_eyetrack_raw.R
\name{json_eyetrack_raw_events}
\alias{json_eyetrack_raw_events}
\title{json_eyetrack_raw: Generates a json file for the Food Choice derivative database file}
\usage{
json_eyetrack_raw_events()
}
\value{
A json file documenting the Food Choice summary database
}
\description{
This function generates a json file for the Food Choice derivative summary database
}
