% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_ff_baseline.R
\name{json_ff_baseline}
\alias{json_ff_baseline}
\title{json_ff_baseline: Generates a json file for baseline Freddy Fullness ratings from double entry REDCap}
\usage{
json_ff_baseline()
}
\value{
A string with data stored in JSON format containing meta-data for baseline Freddy Fullness ratings
}
\description{
This function generates a json file for baseline Freddy Fullness ratings
}
