% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_ff_v3.R
\name{json_ff_v3}
\alias{json_ff_v3}
\title{json_ff_v3: Generates a json file for follow-up Freddy Fullness ratings from double entry REDCap}
\usage{
json_ff_v3()
}
\value{
A string with data stored in JSON format containing meta-data for follow-up Freddy Fullness ratings
}
\description{
This function generates a json file for follow-up Freddy Fullness ratings
}
