% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_ffq_helix.R
\name{json_ffq_helix}
\alias{json_ffq_helix}
\title{json_ffq_helix: Generates a json file for the HELIX study Food Frequency Questionnaire}
\usage{
json_ffq_helix()
}
\value{
A json file documenting the raw inputs and scored values for the HELIX study Food Frequency Questionnaire
}
\description{
This function generates a json file for the scored HELIX study Food Frequency Questionnaire and raw participant responses. This function provides accurate json files ONLY if data is processed using score_ffq_helix function in dataprepr and is only accurate for data collected in Study BRAKE.
}
