% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_fnirs.R
\name{json_fnirs}
\alias{json_fnirs}
\title{json_fnirs: Generates a json file for information about fnirs}
\usage{
json_fnirs()
}
\value{
A json file documenting the Study BRAKE participant fnirs
}
\description{
This function generates a json file about fnirs
}
