% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_fnirscap.R
\name{json_fnirscap}
\alias{json_fnirscap}
\title{json_fnirscap: Generates a json file for information about fnirs cap and headsize}
\usage{
json_fnirscap()
}
\value{
A json file documenting the Study BRAKE participant fnirs cap and headsize
}
\description{
This function generates a json file about fnirs cap and headsize
}
