% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_fnirsinfo_v3.R
\name{json_fnirsinfo_v3}
\alias{json_fnirsinfo_v3}
\title{json_fnirsinfo_v3: Generates a json file for contextual information about the fnirs task during visit 3/follow-up}
\usage{
json_fnirsinfo_v3()
}
\value{
A json file documenting the Study BRAKE's contextual information about the fnirs taskduring visit 3/follow-up
}
\description{
This function generates a json file for contextual information about the fnirs taskduring visit 3/follow-up
}
