% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_fnirstasks_info.R
\name{json_fnirstasks_info}
\alias{json_fnirstasks_info}
\title{json_fnirstasks_info: Generates a json file for contextual information about the fnirs tasks}
\usage{
json_fnirstasks_info()
}
\value{
A json file documenting the Study BRAKE's contextual information about the fnirs tasks
}
\description{
This function generates a json file for contextual information about the fnirs tasks
}
