% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_foodchoice.R
\name{json_foodchoice}
\alias{json_foodchoice}
\title{json_foodchoice: Generates a json file for the Food Choice derivative database file}
\usage{
json_foodchoice()
}
\value{
A json file documenting the Food Choice summary database
}
\description{
This function generates a json file for the Food Choice derivative summary database
}
