% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_foodchoice_events.R
\name{json_foodchoice_events}
\alias{json_foodchoice_events}
\title{json_foodchoice_events: Generates a json file for the Food Choice events file}
\usage{
json_foodchoice_events()
}
\value{
A json file documenting the individual level Food Choice events files in rawdata
}
\description{
This function generates a json file for the individual level Food Choice events files in rawdata
}
