% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_foodrating_events.R
\name{json_foodrating_events}
\alias{json_foodrating_events}
\title{json_foodrating_events: Generates a json file for the Food Rating Events file}
\usage{
json_foodrating_events()
}
\value{
A json file documenting the individual level Food Rating events files in rawdata
}
\description{
This function generates a json file for the individual level Food Rating events files in rawdata
}
