% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_hfi.R
\name{json_hfi}
\alias{json_hfi}
\title{json_hfi: Generates a json file for the Fulkerson Home Food Inventory}
\usage{
json_hfi()
}
\value{
A json file documenting the raw inputs and scored values for the Fulkerson Home Food Inventory
}
\description{
This function generates a json file for the scored Fulkerson Home Food Inventory and raw participant responses. This function provides accurate json files ONLY if data is processed using score_hfi function in dataprepr and is only accurate for data collected in Study BRAKE.
}
