% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_intake_de.R
\name{json_intake_de}
\alias{json_intake_de}
\title{json_intake_de: Generates a json file for intake data from double entry REDCap}
\usage{
json_intake_de()
}
\value{
A string with data stored in JSON format containing meta-data for intake data
}
\description{
This function generates a json file for intake data
}
