% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_meal_info.R
\name{json_meal_info}
\alias{json_meal_info}
\title{json_meal_info: Generates a json file for meal information, fullness, liking, and wanting}
\usage{
json_meal_info()
}
\value{
A string with data stored in JSON format containing meta-data for meal information, fullness, liking, and wanting
}
\description{
This function generates a json file for meal information, fullness, liking, and wanting
}
