% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_intake_v3.R
\name{json_meal_info_v3}
\alias{json_meal_info_v3}
\title{json_meal_info_v3: Generates a json file for meal information, fullness, liking, and wanting}
\usage{
json_meal_info_v3()
}
\value{
A string with data stored in JSON format containing meta-data for meal information, fullness, liking, and wanting
}
\description{
This function generates a json file for meal information, fullness, liking, and wanting
}
