% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_micro_beh.R, R/json_micro_eah.R
\name{json_micro_beh}
\alias{json_micro_beh}
\title{json_micro_beh: Generates a json file for micro_beh}
\usage{
json_micro_beh()

json_micro_beh()
}
\value{
A json file documenting the Study BRAKE's micro_beh

A json file documenting the Study BRAKE's eah_bites
}
\description{
This function generates a json file for micro_beh

This function generates a json file for eah_bites
}
