% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_micro_events.R
\name{json_micro_events}
\alias{json_micro_events}
\title{json_micro_events: Generates a json file for micro_events}
\usage{
json_micro_events()
}
\value{
A json file documenting the Study BRAKE's micro_events
}
\description{
This function generates a json file for micro_events
}
