% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_nik_hfe.R
\name{json_nki_hfe}
\alias{json_nki_hfe}
\title{json_nki_hfe: Generates a json file for the Home Food Environment Survey from the Neighborhood Impact on Kids Study}
\usage{
json_nki_hfe()
}
\value{
A json file documenting the raw inputs and scored values for the Home Food Environment Survey from the Neighborhood Impact on Kids Study
}
\description{
This function generates a json file for the scored Home Food Environment Survey from the Neighborhood Impact on Kids Study and raw participant responses. This function provides accurate json files ONLY if data is processed using score_nki_hfe function in dataprepr and is only accurate for data collected in Study BRAKE.
}
