% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_nonwear.R
\name{json_nonwear}
\alias{json_nonwear}
\title{json_nonwear: Generates a json file for the Non-Wear logs for baseline actigraphy - NOTE: we did not use at follow-up due to low return/completion}
\usage{
json_nonwear()
}
\value{
A string with data stored in JSON format containing meta-data for the Non-Wear logs for baseline actigraphy
}
\description{
This function generates a json file for the Non-Wear logs for baseline actigraphy
}
