% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_participant_info.R
\name{json_participant_info}
\alias{json_participant_info}
\title{json_participant_info: Generates a json file for general participant info}
\usage{
json_participant_info()
}
\value{
A json file documenting the Study BRAKE participant information in the participant_protocol dataset
}
\description{
This function generates a json file for general participant information and protocol information pertaining to rurality status and task/study status for Study BRAKE.
}
