% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_pds.R
\name{json_pds}
\alias{json_pds}
\title{json_pds: Generates a json file for Parent-Report Pubertal Development Scale}
\usage{
json_pds()
}
\value{
A json file documenting the raw inputs and scored values for the Pubertal Development Score.
}
\description{
This function generates a json file for the scored Parent-Pubertal Development Scale and and raw participant responses. This function provides accurate json files ONLY if data is processed using score_pds function in dataprepr and is only accurate for data collected in Study BRAKE.
}
