% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_pref_rank.R
\name{json_pref_rank}
\alias{json_pref_rank}
\title{json_pref_rank: Generates a json file for Taste-Test foods preference-rank}
\usage{
json_pref_rank()
}
\value{
A json file documenting the Study BRAKE's Taste-Test foods preference-rank
}
\description{
This function generates a json file for Taste-Test foods preference-rank
}
