% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_shapegame.R
\name{json_shapegame}
\alias{json_shapegame}
\title{json_shapegame: Generates a json file for the Shape Game derivative file}
\usage{
json_shapegame()
}
\value{
A json file documenting wide formatted Shape Game database
}
\description{
This function generates a json file for the derivative Shape Game database
}
