% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_shapegame_events.R
\name{json_shapegame_events}
\alias{json_shapegame_events}
\title{json_shapegame_events: Generates a json file for the Shape Game events files}
\usage{
json_shapegame_events()
}
\value{
A json file documenting the individual level Shape Game events files in rawdata
}
\description{
This function generates a json file for the individual level Shape Game events files in rawdata
}
