% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_shapegame_info.R
\name{json_shapegame_info}
\alias{json_shapegame_info}
\title{json_shapegame_info: Generates a json file for contextual information about the shapegame}
\usage{
json_shapegame_info()
}
\value{
A json file documenting the Study BRAKE's contextual information about the shapegame
}
\description{
This function generates a json file for contextual information about the shapegame
}
