% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_shapegame_long.R
\name{json_shapegame_long}
\alias{json_shapegame_long}
\title{json_shapegame_long: Generates a json file for the Shape Game long derivative file}
\usage{
json_shapegame_long()
}
\value{
A json file documenting long formatted Shape Game database
}
\description{
This function generates a json file for the long derivative Shape Game database
}
