% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_sleeplog.R
\name{json_sleeplog}
\alias{json_sleeplog}
\title{json_sleeplog: Generates a json file for the Behavioral Rating Inventory of Executive Function-2}
\usage{
json_sleeplog()
}
\value{
A json file documenting the raw inputs and scored values for the Behavioral Rating Inventory of Executive Function-2
}
\description{
This function generates a json file for the scored Behavioral Rating Inventory of Executive Function-2 and raw participant responses. This function provides accurate json files ONLY if data is processed using score_sleeplog function in dataprepr and is only accurate for data collected in Study BRAKE.
}
