% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_spacegame.R
\name{json_spacegame}
\alias{json_spacegame}
\title{json_spacegame: Generates a json file for the Space Game summary behavioral dataset}
\usage{
json_spacegame()
}
\value{
A json file documenting the Space Game summary behavioral dataset
}
\description{
This function generates a json file for the Space Game summary behavioral dataset
}
