% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_spacegame_dm.R
\name{json_spacegame_dm}
\alias{json_spacegame_dm}
\title{json_spacegame_dm: Generates a json file for the Space Game decision making model summary output from matlab}
\usage{
json_spacegame_dm()
}
\value{
A json file documenting the summary decision making model results derivatives/beh file
}
\description{
This function generates a json file for summary decision making model results save in the derivatives/beh file
}
