% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_tastetest.R
\name{json_tastetest}
\alias{json_tastetest}
\title{json_tastetest: Generates a json file for the Taste-Test derivative database file}
\usage{
json_tastetest()
}
\value{
A json file documenting the Taste-Test derivative database
}
\description{
This function generates a json file for the sTaste-Test derivative database
}
