% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_tastetest_events.R
\name{json_tastetest_events}
\alias{json_tastetest_events}
\title{json_tastetest_events: Generates a json file for the Taste-Test Events file}
\usage{
json_tastetest_events()
}
\value{
A json file documenting the individual level Taste-Test events files in rawdata
}
\description{
This function generates a json file for the individual level Taste-Test events files in rawdata
}
