% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_tastetest_long.R
\name{json_tastetest_long}
\alias{json_tastetest_long}
\title{json_tastetest_long: Generates a json file for the Taste-Test derivative database long-formated file}
\usage{
json_tastetest_long()
}
\value{
A json file documenting the Taste-Test derivative database long-formated file
}
\description{
This function generates a json file for the Taste-Test derivative database long-formated file
}
