% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_tt_food.R
\name{json_tt_food}
\alias{json_tt_food}
\title{json_tt_food: Generates a json file for Taste-test sample data from double entry REDCap}
\usage{
json_tt_food()
}
\value{
A string with data stored in JSON format containing meta-data for Taste-test sample
}
\description{
This function generates a json file for Taste-test sample
}
