% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_v2_info.R
\name{json_v2_info}
\alias{json_v2_info}
\title{json_v2_info: Generates a json file for contextual information about the visit 2 protocol tasks}
\usage{
json_v2_info()
}
\value{
A json file documenting the Study BRAKE's contextual information about the visit 2 protocol tasks
}
\description{
This function generates a json file for contextual information about the visit 2 protocol tasks
}
