% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_wasi.R
\name{json_wasi}
\alias{json_wasi}
\title{json_wasi: Generates a json file for WASI}
\usage{
json_wasi()
}
\value{
A json file documenting the Study BRAKE's WASI
}
\description{
This function generates a json file for WASI
}
