% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_task_derivs.R
\name{proc_task_derivs}
\alias{proc_task_derivs}
\title{proc_task_derivs: Generate derivative task data from individual files in rawdata}
\usage{
proc_task_derivs(
  base_wd,
  overwrite = FALSE,
  fnirs_overwrite = FALSE,
  task_list = "all",
  return_data = FALSE
)
}
\arguments{
\item{base_wd}{full path to directory containing both raw_untouched and bids directories}

\item{overwrite}{logical indicating if data should be overwritten. Default = FALSE}

\item{fnirs_overwrite}{overwrite fNIRS data files - separate argument because the Matlab fNIRS processing writes onset values to the rawdata/*_events.tsv values. Overwriting will overwrite these values with NA.}

\item{task_list}{tasks to process. Options include 'all' to process all task data or a list of the following:\itemize{
 \item{'foodrating' - fNIRS Food Rating task}
 \item{'foodchoice' - fNIRS Food Choice task}
 \item{'shapegame' - Shape Game data}
 \item{'spacegame' - Space Game data (need to finish processing in Matlab)}
 \item{'nihtoolbox' - NIH Toolbox data}
 \item{'tastetest' - fNIRS Taste-Test task}
 \item{'pit' - Pavlovian Instrumental Transfer task data}
}}

\item{proc_source}{whether to processes the raw data using proc_tasks.R. Default = FALSE because this takes a long time}
}
\description{
This function: \itemize{
\item{1) processes task data from rawdata to generate derivative data in bids/derivatives}
\item{2) generate json files for derivative task files}
\item{3) write data and json files}
}
}
\examples{

\dontrun{
# organize task data for space game and NIH toolbox in untouchedRaw into sourcedata and raw data
proc_task(base_wd = base_wd, task_list = c('spacegame', 'nih_toolbox'))

}


}
