% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_demo_restricted.R
\name{score_demo_restricted}
\alias{score_demo_restricted}
\title{score_demo_restricted: Score restricted demographic data from Study BRAKE}
\usage{
score_demo_restricted(demo_data, score_base = TRUE, id)
}
\arguments{
\item{demo_data}{a data.frame with restricted PPI information from Study BRAKE (from util_redcap_prepost1.R)}

\item{score_base}{the smallest value assigned to a choice is 0 (i.e., range 0-3). Default = TRUE.}

\item{id}{(optional) name of participant ID column in bes_data. If included the output dataset will be matched by id, if not included the output dataset will be in the order of bes_data but will have no participant identifier. Required to get the phenotype dataset (raw data merged with scores.)}
}
\value{
A dataset with summary demographic information for each participant
}
\description{
This function scores restricted demographic data from Study BRAKE and provides summary information that is less specific/restricted
}
\details{
Note, as long as variable names match those listed, the dataset can include other variables
}
\examples{

# scoring for the hfi with IDs
demo_data <- score_demo_restricted(demo_data, id = 'ID')


}
