% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_actigraph_org_sourcedata.R
\name{util_actigraph_org_sourcedata}
\alias{util_actigraph_org_sourcedata}
\title{util_actigraph_org_sourcedata: Organize actigraphy raw data into BIDS format sourcedata}
\usage{
util_actigraph_org_sourcedata(
  sub_str,
  ses,
  dir_name,
  base_wd,
  overwrite = FALSE
)
}
\arguments{
\item{sub_str}{participant string (i.e., sub_###) as character}

\item{ses}{session name (i.e., 'baseline', 'followup')}

\item{dir_name}{directory name for raw actigraph data (baseline: 'actigraphy'; followup = 'actigraphy_v3')}

\item{base_wd}{absolute path to the data directory}

\item{overwrite}{logical indicating if data should be overwritten. Default = FALSE}
}
\description{
This function copies data from raw_untouched and saves it in sourcedata
}
\details{
To use this function, the correct path must be used. The path must be the full path to the data file, including the participant number.
}
\examples{

# organize actigraphy data
org_actigraph <- util_actigraph_org_sourcedata(tsub_str = 'sub_001', ses = 'baseline', base_wd, overwrite = TRUE)

\dontrun{
}

}
