% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_foodchoice_summary.R
\name{util_foodchoice_summary}
\alias{util_foodchoice_summary}
\title{util_foodchoice_summary: Get summary data from individual participants for the Food Choice task}
\usage{
util_foodchoice_summary(ind_data)
}
\arguments{
\item{ind_data}{Processed individual dataset from rawdata for the Food Choice task}
}
\value{
a data.frame with 1 row including summary performance and task metrics for a participant
}
\description{
This function generates summary performance data for an individual participant
}
\examples{

# process task data for the Food Choice Task
foodchoice_summary_dat <- util_foodchoice_summary(ind_data)

\dontrun{
}


}
