% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_foodrating_summary.R
\name{util_foodrating_summary}
\alias{util_foodrating_summary}
\title{util_foodrating_summary: Get summary data from individual participants for the Food Rating task}
\usage{
util_foodrating_summary(ind_data)
}
\arguments{
\item{ind_data}{Processed individual dataset from rawdata for the Food Rating task}
}
\value{
a data.frame with 1 row including summary performance and task metrics for a participant
}
\description{
This function calculates summary performance data for an individual participant
}
\examples{

# process task data for the Food Choice Task
foodrating_summary_beh <- util_foodrating_summary(ind_data)

\dontrun{
}


}
