% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_format_cwc_data.R
\name{util_format_cwc_data}
\alias{util_format_cwc_data}
\title{util_format_cwc_data: process Child Weight Concerns Questionnaire data}
\usage{
util_format_cwc_data(cwc_data)
}
\arguments{
\item{cwc_data}{rank extracted from data from REDCap events}
}
\description{
This function process Child Weight Concerns Questionnaire data
}
\examples{

# process data
cwc_data_formatted <- util_format_cwc_data(cwc_data)

}
\seealso{
[util_redcap_chile_v5()]
}
