% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_format_ffq_data.R
\name{util_format_ffq_data}
\alias{util_format_ffq_data}
\title{util_format_ffq_data: Prepare Helix FFQ data for scoring}
\usage{
util_format_ffq_data(ffq_data)
}
\arguments{
\item{ffq_data}{ffq data (Helix Food Frequency Questionnaire) extracted from data from REDCap events}
}
\description{
This function prepares Helix FFQ data for scoring with dataprepr::score_ffq_helix()
}
\examples{

# process ffq data
ffq_data_formatted <- util_format_ffq_data(ffq_data)

}
\seealso{
[util_redcap_parent1()], [util_redcap_parent3()]
}
