% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_format_household_data.R
\name{util_format_household_data}
\alias{util_format_household_data}
\title{util_format_household_data: process household data for Study BRAKE visit 1 and 5}
\usage{
util_format_household_data(household_data)
}
\arguments{
\item{household_data}{Household extracted from data from REDCap events}
}
\description{
This function process household data
}
\examples{

# process data
household_data_formatted <- util_format_household_data(household_data)

}
\seealso{
[util_redcap_parent1()], [util_redcap_parent3()]
}
