% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_format_puberty_data.R
\name{util_format_puberty_data}
\alias{util_format_puberty_data}
\title{util_format_puberty_data: Prepare parent-reported puberty data for scoring}
\usage{
util_format_puberty_data(puberty_data)
}
\arguments{
\item{puberty_data}{Puberty data (Parental Rating Scale for Pubertal Development and Tanner) extracted from data from REDCap events}
}
\description{
This function prepares parent-reported puberty data for scoring with dataprepr::score_puberty()
}
\examples{

# process puberty data
puberty_data_formatted <- util_format_puberty_data(puberty_data, respondent = 'child')

}
\seealso{
[util_redcap_parent1()], [util_redcap_parente()]
}
