% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_group_foodrating.R
\name{util_group_foodrating}
\alias{util_group_foodrating}
\title{util_group_foodrating: Get summary data from the Food Rating task}
\usage{
util_group_foodrating(
  data_list,
  ses,
  base_wd,
  overwrite = FALSE,
  return_data = FALSE
)
}
\arguments{
\item{data_list}{A data frame with variable 'sub_str' that includes all participants that have task data in rawdata}

\item{ses}{session name (i.e., 'baseline', 'followup')}

\item{base_wd}{absolute path to the data directory}

\item{overwrite}{logical indicating if data should be overwritten. Default = FALSE}
}
\value{
If return_data is set to TRUE, will return a list including a clean raw dataset with meta-data
}
\description{
This function calculates summary performance data and saves the output
}
\details{
To use this function, the correct path must be used. The path must be the full path to the data file, including the participant number.
}
\examples{

# process task data for the Food Rating Task
group_foodrating_data <- util_group_foodrating(data_list, ses, base_wd, return = TRUE)

\dontrun{
}


}
