% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_merge_questionnaires.R
\name{util_merge_questionnaires}
\alias{util_merge_questionnaires}
\title{util_merge_questionnaires: merges data for questionnaires that were repeated at baseline and follow-up}
\usage{
util_merge_questionnaires(
  child_v1_data,
  child_v2_data,
  child_v3_data,
  parent_v1_data,
  parent_v2_data,
  parent_v3_data
)
}
\arguments{
\item{child_v1_data}{full set of child visit 1 data generated by util_redcap_child_v1}

\item{child_v2_data}{full set of child visit 2 data generated by util_redcap_child_v4}

\item{child_v3_data}{full set of child visit 3 data generated by util_redcap_child_v5}

\item{parent_v1_data}{full set of parent visit 1 data generated by util_redcap_parent_v1}

\item{parent_v2_data}{full set of parent visit 2 data generated by util_redcap_parent_v2}

\item{parent_v3_data}{full set of parent visit 3 data generated by util_redcap_parent_v3}
}
\value{
Will return a list of merged data files:
 \itemize{
   \item{hfi_all}
   \item{loc_all}
   \item{sic_all}
   \item{household_all}
   \item{cebq_all}
   \item{cbq_all}
   \item{cshq_all}
   \item{pstca_all}
   \item{audit_all}
   \item{pmum_all}
   \item{cfpq_all}
   \item{rank_all}
   \item{class_all}
   \item{puberty_all}
   \item{updates_all}
 }
}
\description{
This function data for questionnaires that were repeated at baseline and follow-up. The following questionnaires will be merged across visits and returned individually in a list:
}
\examples{

# process data
merged_qs_list <- util_merge_questionnaires(child_v1_data, child_v2_data, child_v3_data, parent_v1_data, parent_v2_data, parent_v3_data)

}
\seealso{
[proc_redcap(), util_redcap_child1(), util_redcap_child2(),util_redcap_child3(), util_redcap_parent_v1(), util_redcap_parent_v2(), util_redcap_parent_v3()]
}
