% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_merge_anthro.R
\name{util_merged_anthro}
\alias{util_merged_anthro}
\title{util_merged_anthro: merges and formats anthropometrics data}
\usage{
util_merged_anthro(visit1_anthro, visit3_anthro, household_all, date_data)
}
\arguments{
\item{visit1_anthro}{measured anthropometrics at visit 1 from util_redcap_child1()}

\item{visit3_anthro}{measured anthropometrics at visit 3 from util_redcap_child3()}

\item{household_all}{merged household data.frame from merged_qs_list()}
}
\description{
This function merges anthropometrics data across visits and formats/calculates necessary values
}
\examples{

# process data
merge_anthro_data <- util_merged_anthro(bodpod_data = proc_de_data$bodpod_data$data, household_all = merged_qs_list$household_all$data, date_data = date_data)

}
\seealso{
[proc_redcap(), util_redcap_child1(), util_redcap_child3(), util_merge_questionnaires()]
}
