% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_merge_demo.R
\name{util_merged_demo}
\alias{util_merged_demo}
\title{util_merged_demo: merges and formats demographic data}
\usage{
util_merged_demo(visit1_demo, household_all, merged_anthro, date_data)
}
\arguments{
\item{visit1_demo}{visit 1 demo data.frame}

\item{household_all}{merged household data.frame}

\item{merged_anthro}{merged anthro data.fram}
}
\description{
This function merges demographic data across visits and formats/calculates necessary values
}
\examples{

# process data
merge_demo_data <- util_merged_demo(visit1_demo = parent_v1_data$demo_data$data, household_all = merged_qs_list$household_all$data, merged_anthro = merged_anthro$data, date_data = date_data)

}
\seealso{
[proc_redcap(), util_redcap_parent1(), util_redcap_de(), util_merge_questionnaires()]
}
