% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_merge_fnirs.R
\name{util_merged_fnirs}
\alias{util_merged_fnirs}
\title{util_merged_fnirs: merges and formats all fNIRS related data for BIDS}
\usage{
util_merged_fnirs(child_v1_data, child_v3_data, proc_de_data)
}
\arguments{
\item{child_v1_data}{full set of child_v1_data generated by util_redcap_child1}

\item{child_v3_data}{full set of child_v1_data generated by util_redcap_child1}

\item{proc_de_data}{full set of processed double-entry data generated by util_redcap_de}
}
\description{
This function merges all fNIRS related data across visits
}
\examples{

# process data
merged_participants <- util_merged_fnirs(child_v1_data, child_v3_data, proc_de_data)

}
\seealso{
[proc_redcap()], [util_merge_demo()]
}
