% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_merge_intake.R
\name{util_merged_intake}
\alias{util_merged_intake}
\title{util_merged_intake: merges and formats all data related to meals and EAH}
\usage{
util_merged_intake(child_v1_data, child_v3_data, proc_de_data, tastetest_data)
}
\arguments{
\item{child_v1_data}{full set of child_v1_data generated by util_redcap_child1}

\item{child_v3_data}{full set of child_v1_data generated by util_redcap_child1}

\item{proc_de_data}{full set of processed double-entry data generated by util_redcap_de}

\item{tastetest_data}{(optional - needed only for intake data) processed Taste-Test data from fNIRS task so can add liking data to intake data}
}
\description{
This function merges all data related to meals and EAH and calculates intake values
}
\examples{

# process data
merged_intake <- util_merged_intake(child_v1_data, child_v3_data, proc_de_data, tastetest_data)

}
\seealso{
[proc_redcap(), util_redcap_child1(), util_redcap_child3(), proc_task_derivs(), util_group_tastetest(), util_redcap_de(), util_calc_intake()]
}
