% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_merge_participants.R
\name{util_merged_participants}
\alias{util_merged_participants}
\title{util_merged_participants: merges and formats participants data for BIDS}
\usage{
util_merged_participants(merged_demo, date_data)
}
\arguments{
\item{merged_demo}{merged demo data.frame}
}
\description{
This function merges demographic data across visits and formats/calculates necessary values to make the participants.tsv BIDS file
}
\examples{

# process data
merged_participants <- util_merged_participants(merged_demo = merged_demo$data, date_data)

}
\seealso{
[proc_redcap()], [util_merge_demo()]
}
