% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_microstructure.R
\name{util_microstructure}
\alias{util_microstructure}
\title{util_microstructure: Process raw microstructure coding data}
\usage{
util_microstructure(base_wd, file_str, paradigm, ses_str)
}
\arguments{
\item{base_wd}{full path to directory containing both raw_untouched and bids directories}

\item{file_str}{full path to exported ObserverXT coded data}

\item{paradigm}{eating paradigm entered as string. Options include: 'meal' or 'eah'}
}
\value{
Will return a list including data and metadata for
#' \itemize{
 \item{'beh_wide' - microstructure summary metrics}
 \item{'event' - full event-based data for each participant (long format)}
}
}
\description{
This function loads the .txt raw data files from ObserverXT. Cleaning the data involves:
\itemize{
   \item{ 1) loading data and cleaning data (fixing names, getting coder info, unique food bites/food combos)}
   \item{ 2) splitting coders and generating a wide summary microstructure data frame}
   \item{ 3) coding switches and foods tried}
   \item{ 4) computing all summary data}
}
}
\examples{
#if in same working directory as data:
microstructure_list <- util_microstructure(base_wd_path)

\dontrun{

}


}
