% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_nirs_demo.R
\name{util_nirs_demo}
\alias{util_nirs_demo}
\title{util_nirs_demo: Compile and save demographic information of interest for fNIRS analyses}
\usage{
util_nirs_demo(
  v1_demo_homeloc,
  fnirs_info,
  anthro_data,
  followup_anthro_data,
  demographics,
  puberty,
  bodpod,
  baseline_cams,
  followup_cams,
  fullness_tastetest,
  v3_date,
  return_data = TRUE
)
}
\arguments{
\item{v1_demo_homeloc}{demo data from generated from util_redcap_prepost1.R}

\item{fnirs_info}{fnirs task information generated from util_redcap_child1.R}

\item{anthro_data}{anthropometric data from visit 1 generated from util_redcap_child1.R}

\item{followup_anthro_data}{anthropometric data from visit 3 generated from util_redcap_child3.R}

\item{demographics}{demographic data from visit 1 survey generated from util_redcap_parent1.R}

\item{puberty}{parent-report puberty data visit 1 survey generated from util_redcap_parent1.R}

\item{bodpod}{double-entered bodpod generated from util_redcap_de.R}

\item{baseline_cams}{double-entered baseline CAMS generated from util_redcap_de.R}

\item{followup_cams}{double-entered follow-up CAMS generated from util_redcap_de.R}

\item{fullness_tastetest}{double-entered freddy fullness for taste-test generated from util_redcap_de.R}

\item{v3_date}{date from prepost_v3_data}
}
\value{
If return_data is set to TRUE, will return a list including a clean raw dataset with meta-data
}
\description{
This function extracts demographic information of interest for fNIRS analyses.
}
\details{
To use this function, util_redcap_*.R must be completed so all REDCap demographic information is available.
}
\examples{

# process task data for the Food Choice Task
nirs_demo <- util_nirs_demo(sub_str, ses, data_path, return = TRUE)

\dontrun{
}


}
