% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_openneuro.R
\name{util_openneuro}
\alias{util_openneuro}
\title{util_openneuro: Copy BIDS compliant fNIRS files into open-neuro directory for easier upload}
\usage{
util_openneuro(base_wd, overwrite = FALSE)
}
\arguments{
\item{base_wd}{full path to directory containing both raw_untouched and bids directories}

\item{overwrite}{logical indicating if data should be overwritten. Default = FALSE}
}
\description{
This function: \itemize{
\item{1) copies data from data/bids to data/open-neuro}
\item{2) ensures json files are updated}
}
}
\examples{

\dontrun{
# organize task data for space game and NIH toolbox in untouchedRaw into sourcedata and raw data
util_openneuro(base_wd = base_wd)

}


}
