% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_redcap_child1.R
\name{util_redcap_child1}
\alias{util_redcap_child1}
\title{util_redcap_child1: Organize child visit 1 data from REDCap}
\usage{
util_redcap_child1(data, date_data)
}
\arguments{
\item{data}{data from REDCap event child_visit_1_arm_1}

\item{data_data}{date data generated by util_redcap_dates.R}
}
\value{
Will return a list including:
\itemize{
 \item{clean raw child visit 1 datasets}
 \item{meta-data formated as json for each dataset}
 }

 Returned data includes:
 \itemize{
   \item{fnirs_cap}
   \item{anthro_data}
   \item{shapegame_info}
   \item{fnirs_tasks_info}
   \item{meal_info}
   \item{toolbox_info}
   \item{sleep_wk_data}
   \item{hfi_data}
 }
}
\description{
This function organizes REDCap data from REDCap visit data, event child_visit_1_arm_1
}
\examples{

# process REDCap data
child_visit1_list <- util_redcap_child1(data)

\dontrun{
}

}
\seealso{
[proc_redcap(), util_redcap_dates()]
}
