% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_redcap_child3.R
\name{util_redcap_child3}
\alias{util_redcap_child3}
\title{util_redcap_child3: Organize child visit 3 data from REDCap}
\usage{
util_redcap_child3(data, date_data)
}
\arguments{
\item{data}{data from REDCap event child_visit_3_arm_1}
}
\value{
Will return a list including:
\itemize{
 \item{clean raw child visit 3 datasets}
 \item{meta-data formated as json for each dataset}
 }

 Returned data includes:
 \itemize{
   \item{fnirs_cap}
   \item{anthro_data}
   \item{meal_info}
   \item{fnirs_tasks_info}
   \item{pref_rank}
   \item{loc_data}
   \item{sic_data}
   \item{cwc_data}
   \item{body_es_data}
 }
}
\description{
This function organizes REDCap data from REDCap visit data, event child_visit_3_arm_1
}
\examples{

# process REDCap data
child_visit3_list <- util_redcap_child3(data)

\dontrun{
}

}
\seealso{
[proc_redcap(), util_redcap_dates()]
}
