% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_redcap_de.R
\name{util_redcap_de}
\alias{util_redcap_de}
\title{util_redcap_de: Organize double-entry data from REDCap}
\usage{
util_redcap_de(data, return_data = TRUE)
}
\arguments{
\item{data}{data from REDCap data double entry}

\item{return_data}{return organized data (default = TRUE)}
}
\value{
If return_data is set to TRUE, will return a list including:
 1) double-entered raw data across all visits
 2) meta-data/.json for each dataset
}
\description{
This function organizes double-entry REDCap data for all visits
}
\examples{

# process REDCap data
data_de_proc <- util_redcap_de(data)

\dontrun{
}


}
