% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_redcap_parent1.R
\name{util_redcap_parent1}
\alias{util_redcap_parent1}
\title{util_redcap_parent1: Organize parent visit 1 data from REDCap}
\usage{
util_redcap_parent1(data, date_data)
}
\arguments{
\item{data}{data from REDCap event parent_visit_1_arm_1}
}
\value{
Will return a list including:
\itemize{
 \item{clean raw parent visit 1 datasets}
 \item{meta-data formated as json for each dataset}
 }

 Returned data includes:
 \itemize{
   \item{demo_data}
   \item{infancy_data}
   \item{household_data}
   \item{puberty_data}
   \item{cfq_data}
   \item{cebq_data}
   \item{lbc_data}
   \item{lbc_data}
   \item{brief_data}
   \item{ffq_data}
 }
}
\description{
This function organizes REDCap data from REDCap visit data, event parent_visit_1_arm_1
}
\examples{

# process REDCap data
parent_visit1_list <- util_redcap_parent1(data)

\dontrun{
}

}
\seealso{
[proc_redcap(), util_redcap_dates()]
}
