% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_redcap_parent2.R
\name{util_redcap_parent2}
\alias{util_redcap_parent2}
\title{util_redcap_parent2: Organize parent visit 2 data from REDCap}
\usage{
util_redcap_parent2(data, date_data)
}
\arguments{
\item{data}{data from REDCap event parent_visit_2_arm_1}
}
\value{
Will return a list including:
\itemize{
 \item{clean raw parent visit 2 datasets}
 \item{meta-data formated as json for each dataset}
 }

 Returned data includes:
 \itemize{
   \item{cshq_data}
   \item{bes_data}
   \item{ffbs_data}
   \item{hfe_data}
   \item{spsrq_data}
   \item{cbq_data}
   \item{pwlb_data}
   \item{scpf_data}
   \item{fmcb_data}
   \item{tfeq_data}
 }
}
\description{
This function organizes REDCap data from REDCap visit data, event parent_visit_2_arm_1
}
\examples{

# process REDCap data
parent_visit2_list <- util_redcap_parent2(data)

\dontrun{
}

}
\seealso{
[proc_redcap(), util_redcap_dates()]
}
