% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_redcap_parent3.R
\name{util_redcap_parent3}
\alias{util_redcap_parent3}
\title{util_redcap_parent3: Organize parent visit 1 data from REDCap (called within proc_redcap.R)}
\usage{
util_redcap_parent3(data, return_data = TRUE)
}
\arguments{
\item{data}{data from REDCap event parent_visit_1_arm_1}

\item{return_data}{return organized data (default = TRUE)}
}
\value{
If return_data is set to TRUE, will return a list including:
 1) clean raw parent 1 datasets
 2) meta-data/.json for each dataset
}
\description{
This function organizes REDCap data from REDCap visit data, event parent_visit_1_arm_1
}
\examples{

# process REDCap data
parent_visit3_data <- util_redcap_parent3(data)

\dontrun{
}


}
