% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_redcap_prepost.R
\name{util_redcap_prepost}
\alias{util_redcap_prepost}
\title{util_redcap_prepost: Organize Pre/Post visits 1-3 data from REDCap}
\usage{
util_redcap_prepost(v1_data, v2_data, v3_data, v3_child_data)
}
\arguments{
\item{v1_data}{data from REDCap events visit_1_prepost_arm_1}

\item{v2_data}{data from REDCap events visit_2_prepost_arm_1}

\item{v3_data}{data from REDCap events visit_3_child_arm_1}
}
\value{
Will return a list including:
\itemize{
 \item{clean raw protocol dataset for visit 1}
 \item{meta-data formated as json for each dataset}
 }

 Returned data includes:
 \itemize{
   \item{restricted - identifiable data}
   \item{particiant_info - general inforamtion about partiicpants and general protocol}
 }
}
\description{
This function organizes REDCap data from event visit_1_prepost_arm_1, visit_2_prepost_arm_1, visit_3_prepost_arm_1
}
\examples{

# process REDCap data
prepost_data <- util_redcap_prepost(data)

\dontrun{
}

}
\seealso{
[proc_redcap()]
}
