% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_redcap_prepost1.R
\name{util_redcap_prepost1}
\alias{util_redcap_prepost1}
\title{util_redcap_prepost1: Organize Pre/Post visit 1 data from REDCap (called within proc_redcap.R)}
\usage{
util_redcap_prepost1(data, return_data = TRUE)
}
\arguments{
\item{data}{data from REDCap event visit_1_prepost_arm_1}

\item{return_data}{return organized data (default = TRUE)}
}
\value{
If return_data is set to TRUE, will return a list including:
 1) clean raw Pre/Post visit 1 datasets
 2) meta-data/.json for each dataset
}
\description{
This function organizes REDCap data from event visit_1_prepost_arm_1
}
\examples{

# process REDCap data
prepost_visit1_data <- util_redcap_prepost1(data)

\dontrun{
}


}
