% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_scholarsphere.R
\name{util_scholarsphere}
\alias{util_scholarsphere}
\title{util_scholarsphere: Copy BIDS data files and de-identify for the purpose of sharing publically.}
\usage{
util_scholarsphere(base_wd, overwrite = FALSE, data_list = "all")
}
\arguments{
\item{base_wd}{full path to directory containing both raw_untouched and bids directories}

\item{overwrite}{logical indicating if data should be overwritten. Default = FALSE}

\item{data_list}{data to process. Options include 'all' to process all data that will be put on ScholarSphere or a list of the following:\itemize{
 \item{'phenotype' - all phenotype informaiton}
 \item{'shapegame' - Shape Game data}
 \item{'spacegame' - Space Game data (need to finish processing in Matlab)}
 \item{'pit' - Pavlovian Instrumental Transfer task data}
}}
}
\description{

}
\examples{

\dontrun{
# organize task data for space game and NIH toolbox in untouchedRaw into sourcedata and raw data
util_scholarsphere(base_wd = base_wd, task_list = c('spacegame', 'pit'))

}


}
