% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_shapegame_eyetrack_beh.R
\name{util_shapegame_eyetrack_beh}
\alias{util_shapegame_eyetrack_beh}
\title{util_shapegame_eyetrack_beh: Get summary data related to eyetracking for a single Shape Game trial for a participant}
\usage{
util_shapegame_eyetrack_beh(dat_eye_trial)
}
\arguments{
\item{dat_eye_trial}{A data frame with variable 'sub_str' that includes all participants}
}
\value{
a dataframe with eye-tracking related summary behavior for a trial of the Shape Game
}
\description{
This function calculates summary performance data for a participant and saves the output in a wide format (overall task) and long format (by block)
}
\details{
To use this function, the correct path must be used. The path must be the full path to the data file, including the participant number.
}
