% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_shapegame_summary.R
\name{util_shapegame_summary}
\alias{util_shapegame_summary}
\title{util_shapegame_summary: Get summary data from individual participants for the Shape Game (Value-Modulated Attentional Capture Task)}
\usage{
util_shapegame_summary(ind_data, format)
}
\arguments{
\item{ind_data}{Processed individual-level dataset from rawdata for the Food Choice task}

\item{format}{Which format to process the data in - options: 'wide' or 'long': \itemize{
\item{wide: summary metrics computed across last 3 blocks to only average across established performance/remove learning}
\item{long: summary metrics computed for block data provided and then compiled in util_group_shapegame.R}
}}
}
\value{
a data.frame with 1 row including summary performance and task metrics for a participant
}
\description{
This function calculates summary performance for individual participants data and saves the output in a wide format (overall task) and long format (by block)
}
\examples{

# process task data for the Food Choice Task
shapegame_summary_data <- util_shapegame_summary(ind_data, format)

\dontrun{
}


}
