% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_summary_spacegame.R
\name{util_summary_spacegame}
\alias{util_summary_spacegame}
\title{util_summary_spacegame: Process cleaned data from the Space Game (2-stage reinforcement learning task) to generate summary metrics}
\usage{
util_summary_spacegame(
  sub_str,
  ses,
  base_wd,
  overwrite = FALSE,
  return_data = FALSE
)
}
\arguments{
\item{sub_str}{participant string (i.e., sub_###) as character}

\item{ses}{session name (i.e., 'baseline', 'followup')}

\item{base_wd}{absolute path to the data directory}

\item{overwrite}{logical indicating if data should be overwritten. Default = FALSE}
}
\value{
If return_data is set to TRUE, will return a list including a clean raw dataset with meta-data
}
\description{
This function generates summary metrics that can be compiled (e.g., sapply function) to create a behavioral database
}
\details{
To use this function, the correct path must be used. The path must be the full path to the data file, including the participant number.
}
\examples{

# process task data for the Food Choice Task
spacegame_task_pardat <- util_summary_spacegame(sub_str, ses, data_path, return = TRUE)

\dontrun{
}


}
