% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_task_eyetrack.R
\name{util_task_eyetrack}
\alias{util_task_eyetrack}
\title{util_task_eyetrack: Process raw eyetracking .hdf5 files}
\usage{
util_task_eyetrack(task_str, sub_str, ses, base_wd, overwrite = FALSE)
}
\arguments{
\item{task_str}{string for task}

\item{sub_str}{participant string (i.e., sub_###) as character}

\item{ses}{session name (i.e., 'baseline', 'followup')}

\item{base_wd}{absolute path to the data directory}

\item{overwrite}{logical indicating if data should be overwritten. Default = FALSE}
}
\value{
If return_data is set to TRUE, will return a list including a clean raw dataset with meta-data
}
\description{
This function: \itemize{
\item{1) converts data to save in BIDS format in rawdata}
\item{2) cleans data to ensure it is deidentified}
}
}
\details{
To use this function, the correct path must be used. The path must be the full path to the data file, including the participant number.
}
\examples{

# process task data for the Food Choice Task
foodchoice_eye_pardat <- util_task_eyetrack(task_str, sub_str, ses, data_path)

\dontrun{
}

}
