% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_task_foodrating.R
\name{util_task_foodrating}
\alias{util_task_foodrating}
\title{util_task_foodrating: Process raw data from the Food Rating Task}
\usage{
util_task_foodrating(
  sub = sub,
  data_path = data_path,
  overwrite = FALSE,
  return_data = FALSE
)
}
\arguments{
\item{sub}{participant number}

\item{data_path}{data_path absolute path to untouched raw task data file}

\item{overwrite}{logical indicating if data should be overwritten. Default = FALSE}

\item{return_data}{logical indicating if data should be returned. Default = FALSE}
}
\value{
If return_data is set to TRUE, will return a list including a clean raw dataset with meta-data
}
\description{
This function 1) copies data from raw_untouched and saves it in sourcedata and then 2) cleans data to save in BIDS format in rawdata
}
\details{
To use this function, the correct path must be used. The path must be the full path to the data file, including the participant number.
}
\examples{

# process task data for the Food Choice Task
foodrating_task_pardat <- util_task_foodrating(data_path, return = TRUE)

\dontrun{
}


}
