% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_task_org_sourcedata.R
\name{util_task_org_sourcedata}
\alias{util_task_org_sourcedata}
\title{util_task_org_sourcedata: Organize task raw data into BIDS format sourcedata}
\usage{
util_task_org_sourcedata(
  task_str,
  sub_str,
  ses,
  base_wd,
  task_cat,
  overwrite = FALSE
)
}
\arguments{
\item{task_str}{string for task}

\item{sub_str}{participant string (i.e., sub_###) as character}

\item{ses}{session name (i.e., 'baseline', 'followup')}

\item{base_wd}{absolute path to the data directory}

\item{task_cat}{bids category/directory for task (e.g., 'nirs', 'beh')}

\item{overwrite}{logical indicating if data should be overwritten. Default = FALSE}
}
\description{
This function copies data from raw_untouched and saves it in sourcedata
}
\details{
To use this function, the correct path must be used. The path must be the full path to the data file, including the participant number.
}
\examples{

# process task data for the Food Choice Task
org_foodchoice <- util_task_org_sourcedata(task_str = 'foodchoice', sub_str = 'sub_001', base_wd, overwrite = TRUE)

\dontrun{
}


}
