% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_task_pit.R
\name{util_task_pit}
\alias{util_task_pit}
\title{util_task_pit: Clean and organize PIT data into BIDS rawdata}
\usage{
util_task_pit(sub_str, ses, base_wd, overwrite = FALSE, return_data = FALSE)
}
\arguments{
\item{sub_str}{participant string (i.e., sub_###) as character}

\item{ses}{session name (i.e., 'baseline', 'followup')}

\item{base_wd}{absolute path to the data directory}

\item{overwrite}{logical indicating if data should be overwritten. Default = FALSE}
}
\value{
statement of task completed
}
\description{
This function formats and organizes RRV data from bids/sourcedata into bids/rawdata for a given subject
}
\examples{

\dontrun{
}
}
