% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_task_spacegame.R
\name{util_task_spacegame}
\alias{util_task_spacegame}
\title{util_task_spacegame: Process raw data from the Space Game (2-stage reinforcement learning task)}
\usage{
util_task_spacegame(
  sub_str,
  ses,
  base_wd,
  overwrite = FALSE,
  return_data = FALSE
)
}
\arguments{
\item{sub_str}{participant string (i.e., sub_###) as character}

\item{ses}{session name (i.e., 'baseline', 'followup')}

\item{base_wd}{absolute path to the data directory}

\item{overwrite}{logical indicating if data should be overwritten. Default = FALSE}
}
\value{
If return_data is set to TRUE, will return a list including a clean raw dataset with meta-data
}
\description{
This function: 1) cleans data to save in BIDS format in rawdata and 2) generates summary data that can be used to generate a database
}
\details{
To use this function, the correct path must be used. The path must be the full path to the data file, including the participant number.
}
\examples{

# process task data for the Food Choice Task
spacegame_task_pardat <- util_task_spacegame(sub_str, ses, data_path, return = TRUE)

\dontrun{
}


}
