% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_tastetest_summary.R
\name{util_tastetest_summary}
\alias{util_tastetest_summary}
\title{util_tastetest_summary: Get summary data from individual participants for the Taste-Test task}
\usage{
util_tastetest_summary(sub_str, raw_wd, nirs_folder, file_name)
}
\arguments{
\item{sub_str}{participant string (i.e., sub_###) as character}

\item{raw_wd}{individual participant rawdata directory path}

\item{nirs_folder}{name of folder within the nirs directory - option: premeal, postmeal}

\item{file_name}{file name of individual *_events.tsv data file to load}
}
\value{
a data.frame with 1 row including summary performance and task metrics for a participant
}
\description{
This function calculates summary performance data for an individual participant
}
\examples{

# process task data for the Food Choice Task
foodrating_summary_beh <- util_tastetest_summary(sub_str, raw_wd, nirs_folder, file_name)

\dontrun{
}


}
