% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_microstructure.R
\name{write_microstructure}
\alias{write_microstructure}
\title{write_microstructure: Write selected data and json files from processed microstructure data}
\usage{
write_microstructure(
  base_wd,
  intake_data,
  overwrite = FALSE,
  micro_protocols = "all",
  micro_data_type = "all",
  return_data = FALSE
)
}
\arguments{
\item{base_wd}{full path to directory containing both raw_untouched and bids directories}

\item{intake_data}{processed intake data from proc_redcap.R}

\item{overwrite}{logical indicating if data should be overwritten. Default = FALSE}

\item{micro_protocols}{list of strings matching the notes below to indicate the which microstructure data. Default = 'all' to export all data and metadata. Options include:
\itemize{
 \item{'meal-baseline' - meal microstructure behavior at baseline}
 \item{'meal-followup' - meal microstructure behavior at followup}
 \item{'eah' - EAH microstructure behavior}
}}

\item{micro_data_type}{Type of data to process for meal microstructure - list of strings matching the data types listed below. Default = 'all' to export both:
\itemize{
  \item{'beh_wide' - summary behavioral measures in wide formate by coder. Note: this will write out a summary dataset in bids/phenotype.}
  \item{'events_long' - event level data in log format by coder. Note: this writes out a file per participant into bids/rawdata.}
}}
}
\value{
Does not return anything
}
\description{
This function:
\itemize{
   \item{1) Calls proc_microstructure function to get clean and compiled data and metadata}
   \item{2) Exports all or select BIDS-compliant .tsv and .json files into bids/phenotype and/or bids/rawdata}
}
}
\details{
To use this function, the correct path must be used. The path must be the full path to the data file, including the file name.
}
\examples{

\dontrun{
write_microstructure(base_wd, intake_data, overwrite = FALSE, micro_protocols = 'all', micro_data_type = 'all')

}

}
