% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataBRAKEr.R
\name{dataBRAKEr}
\alias{dataBRAKEr}
\title{dataBRAKEr: Process Study BRAKE data}
\usage{
dataBRAKEr(
  base_wd,
  overwrite = FALSE,
  fnirs_overwrite = FALSE,
  proc_source = FALSE,
  data_list = "all",
  data_type = "all",
  micro_protocols = "all",
  micro_data_type = "all",
  overwrite_ggir_derivs = FALSE,
  return_data = FALSE
)
}
\arguments{
\item{base_wd}{full path to directory containing both raw_untouched and bids directories}

\item{overwrite}{logical indicating if data should be overwritten. Default = FALSE}

\item{fnirs_overwrite}{overwrite fNIRS data files - separate argument because the Matlab fNIRS processing writes onset values to the rawdata/*_events.tsv values. Overwriting will overwrite these values with NA.}

\item{proc_source}{whether to processes the raw data using proc_tasks.R. Default = FALSE}

\item{data_list}{list of strings matching the notes below to indicate the data to be written. Default = 'all' to export all data and metadata. Options include:
\itemize{
 \item{'foodview' - fMRI Food Viewing task}
 \item{'nihtoolbox' - NIH Toolbox data}
 \item{'pit' - Pavlovian Instrumental Transfer task data}
 \item{'rrv' - Relative Reinforcing Value of Food task}
 \item{'spacegame' - Space Game data (need to finish processing in Matlab)}
 \item{'sst' - fMRI Stop-Signal Task data}
}}

\item{micro_protocols}{(optional) list of strings matching the notes below to indicate the which microstructure data. Default = 'all' to export all data and metadata. Options include:
\itemize{
 \item{'meal-baseline' - meal microstructure behavior at baseline}
 \item{'meal-followup' - meal microstructure behavior at followup}
 \item{'eah' - EAH microstructure behavior}
}}

\item{micro_data_type}{(optional) Type of data to process for meal microstructure - list of strings matching the data types listed below. Default = 'all' to export both:
\itemize{
  \item{'beh_wide' - summary behavioral measures in wide formate by coder. Note: this will write out a summary dataset in bids/phenotype.}
  \item{'events_long' - event level data in log format by coder. Note: this writes out a file per participant into bids/rawdata.}
}}

\item{return_data}{(logical) return data to working environment. Default = FALSE}
}
\value{
Does not return anything
}
\description{
This function:
\itemize{Calls core proc_\* and write_\* functions and writes BIDS-compliant .tsv and .json files into bids/phenotype:
   \item{write_redcap: processes and writes out specified REDCap data}
   \item{proc_microstructure: processes and writes out microstructure data}
   \item{proc_tasks:}
}
}
\details{
To use this function, the correct path must be used. The path must be the full path to the data file, including the file name.
}
\examples{

\dontrun{
dataBRAKEr(base_wd, overwrite = FALSE, data_list = 'all')

}

}
