% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_foodrating.R
\name{json_foodrating}
\alias{json_foodrating}
\title{json_foodrating: Generates a json file for the Food Rating derivative database file}
\usage{
json_foodrating()
}
\value{
A json file documenting the Food Rating derivative database
}
\description{
This function generates a json file for the sFood Rating derivative database
}
