% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_micro_beh.R
\name{json_micro_beh}
\alias{json_micro_beh}
\title{json_micro_beh: Generates a json file for micro_beh}
\usage{
json_micro_beh()
}
\value{
A json file documenting the Study BRAKE's micro_beh
}
\description{
This function generates a json file for micro_beh
}
