% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_redcap.R
\name{proc_redcap}
\alias{proc_redcap}
\title{proc_redcap: Process raw data downloaded from Study BRAKE REDCap}
\usage{
proc_redcap(
  redcap_api = FALSE,
  redcap_visit_data,
  redcap_de_data,
  tastetest_data
)
}
\arguments{
\item{redcap_api}{(logical) execute REDCap API. Default = FALSE.}

\item{redcap_visit_data}{REDCap visit data from a prior API call}

\item{redcap_de_data}{REDCap double-entry data from a prior API call}

\item{tastetest_data}{(optional - needed only for intake data) processed Taste-Test data from fNIRS task so can add liking data to intake data}
}
\value{
Will return a list including data and metadata for:
\itemize{
 \item{'paticipants' - BIDS specified participants.tsv file}
 \item{'anthropometrics' - height, weight, and computed anthropometric data}
 \item{'demographics' - compiled demographic data}
 \item{'bodpod' - verified BodPod data}
 \item{'fnirs_info' - compiled fNIRS task-related information}
 \item{'dkefs' - verified D-KEFS data}
 \item{'wasi' - verified WASI data}
 \item{'intake' - compiled verified intake data with computed intake values}
 \item{'tasttest_samples' - verified Taste-Test sample weights}
 \item{'household' - compiled demographicinformation about houshold}
 \item{'infancy' - compiled demographic information related to infancy}
 \item{'bes' - Binge Eating Scale}
 \item{'brief2' - Behavioral Rating Inventory of Executive Function-2}
 \item{'cbq' - Child Behavior Questionnaire}
 \item{'cebq' - Children's Eating Behavior Questionnaire}
 \item{'cfq' - Child Feeding Questionnaire}
 \item{'cshq' - Children Sleep Habits Questionnaire}
 \item{'efcr' - External Food Cue Responsiveness Scale}
 \item{'ffbs' - Family Food Behavior Survey}
 \item{'ffq' - HELIX cohort Food Frequency Questionnaire}
 \item{'fmcb' - Feeding to Manage Child Behavior Questionnaire}
 \item{'hfe' - Home Food Environment}
 \item{'hfi' - Fulkerson Home Food Inventory}
 \item{'lbc' - Lifestyle Behavior Checklist}
 \item{'loc' - Loss of Control-Eating Questionnaire}
 \item{'puberty' - combination of Tanner and Pubertal Rating Scale}
 \item{'pwlb' - Parent Weight-Loss Behavior Questionnaire}
 \item{'sic' - Stress in Children Questionnaire}
 \item{'sleeplog' - Week long sleep log}
 \item{'spsrq' - Sensitivity to Punishment and Sensitivity to Reward Questionnaire}
 \item{'tfeq' - Three Factor Eating Questionnaire}
}
}
\description{
This function:
\itemize{
   \item{1) Reads REDCap data (visit and double-entry) using the REDCap API}
   \item{2) Calls util_ functions to clean and compile data in dataframes}
   \item{3) Calls json_ functions to create strings with meta-data stored in JSON format for each dataframe}
   \item{4) Compiles data repeated across visits and sessions}
}
}
\details{
To use this function, the correct path must be used. The path must be the full path to the data file, including the file name.
}
\examples{

\dontrun{
redcap_data <- proc_redcap(base_wd, overwrite = FALSE, overwrite_jsons = FALSE)

}

}
\seealso{
[write_redcap()]
}
