% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_redcap.R
\name{proc_redcap}
\alias{proc_redcap}
\title{proc_redcap: Process raw data downloaded from Study BRAKE REDCap}
\usage{
proc_redcap(sourcedata_wd, overwrite = FALSE, return_data = FALSE)
}
\arguments{
\item{sourcedata_wd}{full path to the sourcedata directory}

\item{overwrite}{overwrite existing files (default = FALSE)}

\item{return_data}{return phenotype to console (default = FLASE)}
}
\value{
If return_data is set to TRUE, will return a list including:
 1) clean raw phenotype datasets for each task
 2) meta-data/.json inforamtion for each task
}
\description{
This function:
1) reads REDCap data from sourcedata
2) cleans data to save in BIDS format in phenotype. Produces the following files:
   *
3) calls functions to create .json files for each phoneypte/x.tsv file
}
\details{
To use this function, the correct path must be used. The path must be the full path to the data file, including the file name.
}
\examples{

# process REDCap data
phenotype_data <- proc_redcap(visit_data_path, data_de_path, return = TRUE)

\dontrun{
}


}
