% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_tasks.R
\name{proc_tasks}
\alias{proc_tasks}
\title{proc_tasks: Process raw data from all computer tasks used in Study BRAKE}
\usage{
proc_tasks(data_path, overwrite = FALSE)
}
\arguments{
\item{data_path}{full data path to raw_untouched directory}

\item{overwrite}{logical indicating if data should be overwritten. Default = FALSE}
}
\value{
data.frame for each task with status for each processing step
}
\description{
This function calls task scripts that:
1) copies data from raw_untouched and saves it in sourcedata
2) cleans data to save in BIDS format in rawdata
3) processes rawdata to generate derivative or summary databases for each task
}
\details{
To use this function, the correct path must be used. The path must be the full path to the data file, including the participant number.
}
\examples{

# process task data for the Food Choice Task
proc_tasks_pardat <- proc_tasks(data_path, overwrite)

\dontrun{
}


}
