% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deriv_tasks.R, R/proc_tasks.R
\name{proc_tasks}
\alias{proc_tasks}
\title{deriv_tasks: Generate derivative databases for all computer tasks used in Study BRAKE}
\usage{
proc_tasks(base_wd, overwrite = FALSE)

proc_tasks(base_wd, overwrite = FALSE)
}
\arguments{
\item{base_wd}{full path to directory containing both raw_untouched and bids directories}

\item{overwrite}{logical indicating if data should be overwritten. Default = FALSE}
}
\value{
data.frame for each task with status for each processing step

data.frame for each task with status for each processing step
}
\description{
To use this function, the correct path must be used. The path must be the full path to the data file, including the participant number.

This function calls task scripts that:
1) copies data from raw_untouched and saves it in sourcedata
2) cleans data to save in BIDS format in rawdata
3) processes rawdata to generate derivative or summary databases for each task
}
\details{
To use this function, the correct path must be used. The path must be the full path to the data file, including the participant number.
}
\examples{

# process task data for the Food Choice Task
proc_tasks_pardat <- proc_tasks(base_wd, overwrite)

\dontrun{
}



# process task data for the Food Choice Task
proc_tasks_pardat <- proc_tasks(base_wd, overwrite)

\dontrun{
}


}
