% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_group_shapegame.R
\name{util_group_shapegame}
\alias{util_group_shapegame}
\title{util_group_shapegame: Get summary data from the Shape Game (Value-Modulated Attentional Capture Task)}
\usage{
util_group_shapegame(
  data_list,
  ses,
  base_wd,
  overwrite = FALSE,
  return_data = FALSE
)
}
\arguments{
\item{data_list}{A data frame with variable 'sub_str' that includes all participants}

\item{ses}{session name (i.e., 'baseline', 'followup')}

\item{base_wd}{absolute path to the data directory}

\item{overwrite}{logical indicating if data should be overwritten. Default = FALSE}
}
\value{
If return_data is set to TRUE, will return a list including a clean raw dataset with meta-data
}
\description{
This function calculates summary performance data and saves the output in a wide format (overall task) and long format (by block)
}
\details{
To use this function, the correct path must be used. The path must be the full path to the data file, including the participant number.
}
\examples{

# process task data for the Food Choice Task
group_shapegame_dat <- util_group_shapegame(data_list, ses, base_wd, overwrite, return_data = TRUE)

\dontrun{
}


}
