% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_redcap_de.R
\name{util_redcap_de}
\alias{util_redcap_de}
\title{util_redcap_de: Organize double-entry data from REDCap}
\usage{
util_redcap_de(redcap_api = FALSE, redcap_de_data, date_data)
}
\arguments{
\item{redcap_api}{(logical) execute REDCap API. Default = FALSE.}

\item{redcap_de_data}{REDCap double-entry data from a prior API call}
}
\value{
Will return a list including data that has been double-entered and checked along with the metadata for:
\itemize{
 \item{bodpod}
 \item{wasi}
 \item{dkefs}
 \item{intake}
 \item{fullness_baseline}
 \item{fullness_followup}
 \item{taste_test}
 \item{cams}
}
}
\description{
This function organizes REDCap double entry data data
}
