% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_redcap_de_org.R
\name{util_redcap_de}
\alias{util_redcap_de}
\title{util_redcap_de: Organize REDCap double entry data (called within proc_redcap.R)}
\usage{
util_redcap_de(data, return_data = TRUE)
}
\arguments{
\item{data}{data from REDCap data double entry}

\item{return_data}{return organized data (default = TRUE)}
}
\value{
If return_data is set to TRUE, will return a list including:
 1) clean raw data double entry datasets
 2) meta-data/.json for each dataset
}
\description{
This function organizes REDCap double entry data
}
\examples{

# process REDCap data
double_enter_data <- util_redcap_de(data)

\dontrun{
}


}
