% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_redcap_parent2.R
\name{util_redcap_parent2}
\alias{util_redcap_parent2}
\title{util_redcap_parent2: Organize parent visit 2 data from REDCap (called within proc_redcap.R)}
\usage{
util_redcap_parent2(data, return_data = TRUE)
}
\arguments{
\item{data}{data from REDCap event parent_visit_2_arm_1}

\item{return_data}{return organized data (default = TRUE)}
}
\value{
If return_data is set to TRUE, will return a list including:
 1) clean raw parent visit 2 datasets
 2) meta-data/.json for each dataset
}
\description{
This function organizes REDCap data from REDCap visit data, event parent_visit_2_arm_1
}
\examples{

# process REDCap data
parent_visit2_data <- util_redcap_parent2(data)

\dontrun{
}


}
