% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_redcap_parent3.R
\name{util_redcap_parent3}
\alias{util_redcap_parent3}
\title{#' util_redcap_parent3: Organize parent visit 3 data from REDCap}
\usage{
util_redcap_parent3(data, date_data)
}
\arguments{
\item{data}{data from REDCap event parent_visit_3_arm_1}
}
\value{
Will return a list including:
\itemize{
 \item{clean raw parent visit 3 datasets}
 \item{meta-data formated as json for each dataset}
 }

 Returned data includes:
 \itemize{
   \item{household_data}
   \item{sleep_data}
   \item{hfi_data}
   \item{puberty_data}
   \item{cfq_data}
   \item{efcr_data}
   \item{ffq_data}
   \item{cshq_data}
   \item{bes_data}
   \item{hfe_data}
   \item{pwlb_data}
 }
}
\description{
This function organizes REDCap data from REDCap visit data, event parent_visit_3_arm_1
}
\examples{

# process REDCap data
parent_visit3_list <- util_redcap_parent3(data)

\dontrun{
}

}
\seealso{
[proc_redcap(), util_redcap_dates()]
}
