% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_task_nihtoolbox.R
\name{util_task_nihtoolbox}
\alias{util_task_nihtoolbox}
\title{util_task_nihtoolbox: Process raw data from the NIH Toolbox}
\usage{
util_task_nihtoolbox(
  task,
  sub_str_list,
  base_wd,
  overwrite = FALSE,
  return_data = TRUE
)
}
\arguments{
\item{task}{task string ('flanker', 'dccs', or 'listserve')}

\item{sub_str_list}{list of participant strings for group database}

\item{base_wd}{absolute path to the data directory}

\item{overwrite}{logical indicating if data should be overwritten. Default = FALSE}

\item{return_data}{logical indicating if data should be returned. Default = FALSE}
}
\value{
If return_data is set to TRUE, will return a list including a clean raw dataset with meta-data
}
\description{
This function: 1) cleans data to save in BIDS format in rawdata and 2) generates summary data that can be used to generate a database
}
\details{
To use this function, the correct path must be used. The path must be the full path to the data file, including the participant number.
}
\examples{

# process task data for the Food Choice Task
nihtoolbox_task_pardat <- util_task_nihtoolbox(task, data, base_wd, overwrite, return = TRUE)

\dontrun{
}


}
