% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_redcap.R
\name{write_redcap}
\alias{write_redcap}
\title{write_redcap: Write selected data and json files from processed REDCap data}
\usage{
write_redcap(
  base_wd,
  overwrite = FALSE,
  data_list = "all",
  tastetest_data,
  return_data = FALSE
)
}
\arguments{
\item{base_wd}{full path to directory containing both raw_untouched and bids directories}

\item{overwrite}{logical indicating if data should be overwritten. Default = FALSE}

\item{data_list}{list of strings matching the notes below to indicate the data to be written. Default = 'all' to export all data and metadata. Options include:
\itemize{
 \item{'paticipants' - BIDS specified participants.tsv file}
 \item{'anthropometrics' - height, weight, and computed anthropometric data}
 \item{'demographics' - compiled demographic data}
 \item{'bodpod' - verified BodPod data}
 \item{'fnirs_info' - compiled fNIRS task-related information}
 \item{'dkefs' - verified D-KEFS data}
 \item{'wasi' - verified WASI data}
 \item{'intake' - compiled verified intake data with computed intake values}
 \item{'tasttest_samples' - verified Taste-Test sample weights}
 \item{'household' - compiled demographicinformation about houshold}
 \item{'infancy' - compiled demographic information related to infancy}
 \item{'bes' - Binge Eating Scale}
 \item{'brief2' - Behavioral Rating Inventory of Executive Function-2}
 \item{'cbq' - Child Behavior Questionnaire}
 \item{'cebq' - Children's Eating Behavior Questionnaire}
 \item{'cfq' - Child Feeding Questionnaire}
 \item{'cshq' - Children Sleep Habits Questionnaire}
 \item{'efcr' - External Food Cue Responsiveness Scale}
 \item{'ffbs' - Family Food Behavior Survey}
 \item{'ffq' - HELIX cohort Food Frequency Questionnaire}
 \item{'fmcb' - Feeding to Manage Child Behavior Questionnaire}
 \item{'hfe' - Home Food Environment}
 \item{'hfi' - Fulkerson Home Food Inventory}
 \item{'lbc' - Lifestyle Behavior Checklist}
 \item{'loc' - Loss of Control-Eating Questionnaire}
 \item{'puberty' - combination of Tanner and Pubertal Rating Scale}
 \item{'pwlb' - Parent Weight-Loss Behavior Questionnaire}
 \item{'sic' - Stress in Children Questionnaire}
 \item{'sleeplog' - Week long sleep log}
 \item{'spsrq' - Sensitivity to Punishment and Sensitivity to Reward Questionnaire}
 \item{'tfeq' - Three Factor Eating Questionnaire}
}}

\item{tastetest_data}{(optional - needed only for intake data) processed Taste-Test data from fNIRS task so can add liking data to intake data}
}
\value{
Does not return anything
}
\description{
This function:
\itemize{
   \item{1) Calls proc_redcap function to get clean and compiled data and metadata}
   \item{2) Exports all or select BIDS-compliant .tsv and .json files into bids/phenotype}
}
}
\details{
To use this function, the correct path must be used. The path must be the full path to the data file, including the file name.
}
\examples{

\dontrun{
write_redcap(base_wd, overwrite = FALSE, data_list = 'all')

}


}
